requirements.txt
# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# HTTP Requests
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.2
lxml==5.1.0

# HTML Parsing and Metadata Extraction
html5lib==1.1
extruct==0.16.0

# Database
# sqlite3  # Included in Python standard library

# Environment Variables
python-dotenv==1.0.0

# Date/Time utilities
python-dateutil==2.8.2

# MCP Protocol (Model Context Protocol)
# Nota: Asume que telegram-mcp ya está configurado externamente

# Image Generation API (opcional, según el servicio que uses)
# openai==1.3.0  # Si usas DALL-E
# stability-sdk==0.8.4  # Si usas Stable Diffusion

# Utilities
validators==0.22.0
urllib3==2.1.0